Reworking of Directives
I think we have gotten directives slightly wrong. Since we're still in working draft, I think we should fix this:

Old:

```
{
  foo @if: true
  bar @unless: true
}
```

New:

```
{
  foo @include(if: true)
  bar @skip(if: true)
}
```

Now directives are just as powerful as fields in their ability to accept multiple argument values. More importantly, this is a *simplification* as directive arguments and field arguments are syntactically *identical*, and semantically extremely similar, allowing us to reuse spec, grammar, and impl for both.

This simplification is really valuable to fix some issues with directives today. Directive value nullability is poorly (and probably incorrectly) defined in the current form. What does a non-null type on a directive today mean? Directives cannot be required. Now the rules are the same as field arguments.

This also helps us support future directives that have no arguments, or directives that have multiple arguments. `@defer` alone is perfectly fine, and `@defer(priority: 3)` is much easier to read than `@defer: 3`.

So if we do this, then `@if` and `@unless` are replaced by `@include` and `@skip` respectively. I think this is helpful as it describes more accurately what the directives' jobs are.

So grammar becomes:

```
Directive : Name Arguments?
```

And the directive value validation rule is merged into the argument validation rule.