[RFC] standard printer uses no comma between fields
A couple issues have opened up about inconsistency in rendering graphql queries in the spec and in the ref impl - especially with respect to whitespace and commas.

Here I'm proposing that it's easier to read queries where every field is always on it's own line (eg, no `image { width, height }` one-liners) and that lines are split only by line-terminators and not also commas.

If there's agreement around this proposal, the second step will be to apply the printer to all graphql examples in the spec.