Validation returns list of errors.
This is a small simplification to Validation: rather than returning a struct including both the result of validation as well as a list of errors, it simply returns the list of errors.

Also, importantly, this no longer formats the error within the validation submodule, but instead returns raw GraphQLError objects so that more value can be extracted from these by the caller. Formatting should be done much later in the pipeline.