Simplify coveralls script
Removes `npm run cover` in favor of the one `npm run coveralls` integration.

Also simplifies that script, removing the need for babel-node.

Finally, ensures the root is at "src" which removes the transient error parsing _mocha, speeds up the script, and does not count "scripts" as part of coverage score.