NamedType AST
This replaces the raw usage of `Name` as a kind of `Type` with the AST node `NamedType`. It also carries the vernacular "named type" throughout the codebase in the places where this same concept was being provided many different names (unmodified, naked, raw, etc).

Finally, this simplifies some of the validation rules once this clarification was in place.