Schema Materializer
Adds the schema materializer. This components takes a schema definition
AST and produces a set of type definition objects corresponding to the
schema definitions. This allows a user to specify schema in a far
more literate fashion

The next tool that I'm going to write is one that takes the
schema definition DSL and outputs the result of the introspection
query against that schema. This will be really useful for tests, etc.