[RFC] Do not format errors from graphql()
This changes the implementation of `graphql()` to not format any returned errors. Most importantly, this allows access to the `stack` property of each error, along with richer information contained in GraphQLError like the affected AST nodes and source document.

This means that the responsibility for formatting errors must be handled elsewhere, likely by an HTTP service like `express-graphql`.