[RFC] move type validation to ctor invariants
This removes the type validation system, and in its place puts invariants into the type ctors.

This ensures that errors are produced as types are created, rather than when (or if) they run through the validator. This should lead to faster error recovery and less errors.

There were a number of issues detected in our test files that these invariants picked up.

Fixes #121