Resolvers can return `Error` to signify failure
This adds another way resolver functions can signify a failure: returning an Error object.

The primary reason this is useful is returning a list of values where index contains an error but another does not. A test casehas been added to illustrate this behavior. Previously sync resolvers could only throw an error, killing the whole list, or return nulls and fail to present error messaging.

I then used this as a simplifying factor to the executor core where it removed a third instance of locating and logging an error.