Define possibleTypes on InterfaceType
Mutation of interfaces in type definition is problematic,
as it allows for incomplete schema generation if
all types are not loaded at runtime.

This adds an optional `possibleTypes` property which may be
an Array or thunk, resolving with the possible types an
interface may implement. It also allows a thunk for interfaces
to act as intended, currently it is immediately invoked via
`addImplementationToInterfaces` call in constructor.