Spec compliant Int sizing
As discussed in #182, this was a deviation between spec and reference implementation, where the spec is more reasonable.

This tightens the allowed range of Int values to valid 32-bit signed integers, supporting the broadest collection of platforms.

For those who stumble upon this rev looking for ways to represent numeric-looking large values, see ID or Custom Scalars.