[FIX] Include locations for errors related to failed value completion.
Error locations were previously only being added due to async value resolution and not during value completion. This meant if an error was raised due to an incorrect response from a resolver function that it would not have gotten wrapped in a location-aware GraphQLError object.

This includes a new intermediate phase which catches and assigns these locations before further propogating the error.

This also assigns non-enumable property for `originalError` so it is not included during JSON.stringify.