Use the Error base class stack property
Many libraries (such as https://github.com/PolymerLabs/stacky) expect an error's stack to be of a certain format. Currently this code sets it to the message, which causes exceptions in formatters. 

It also obscures the actual location of where the error is thrown, since it overwrites the stack property of the inherited Error.