Refactor GraphQLError, add many tests
This fixes an issue where `new GraphQLError() instanceof GraphQLError` would unexpectedly return false due to a combination of babel's class transform and the weird behavior of Error on many platforms including Node/v8.

In addition, when the stack from an originating error is not available, this uses Node/v8's Error.captureStackTrace when available for cleaner stack traces.

Finally this adds lots of documentation! Yay!