Update extendSchema to support new Directives (#455)
* Update extendSchema to support new Directives

This changes the extendSchema utility to accept and process new
Directive definitions. Existing directives cannot be extended or
replaced (should throw an error in such cases).

See extendSchema-test for examples.

* Update extendSchema to support new Directives

This changes the extendSchema utility to accept and process new
Directive definitions. Existing directives cannot be extended or
replaced (should throw an error in such cases).

See extendSchema-test for examples.

* Fix trailing whitespace lint error

* Improve logic in new directive extendSchema

Use an Array instead of an object for collecting new
directives detected within the schema extension.