Advantage common convention for naming root types.
This is a result of a conversation I had with @lacker that we should provide a stronger opinion on what to name the root types, and then advantage that opinion. Here's the result of that:

If you use the type names `Query` and `Mutation` for your root types, then printing the schema or parsing a schema IDL can omit the `schema { query: Query, mutation: Mutation }` boilerplate. If your root type names are anything other than these, then you have to include it.

There doesn't seem to be a significant downsite to this except for the relative complication of understanding why sometimes a `schema {}` description is included and sometimes it isn't.