Use linked-list to represent response path.
This is a performance improvement to the execution path that removes the need for creating a new array to represent the response path for each resolved field in the response. Instead this creates a single-linked-list which is naturally persistently immutable.

This exposes a breaking change: if you relied on `resolve(_, _, _, { path }) {` then you'll find `path` in this linked list format instead of an array.