add public aliases for type configs
Having access to the type config is very useful for building type building utilities. Take for instance a function `implementNodeInterface(type: GraphQLObjectType)` which clones the `type._typeConfig` object and adds an interface, verifies an `isTypeOf` field exists, handles a special `nodeResolve` field (which can be used by a query level `node` field), and adds the global `id` field required by the interface.

As utilities around a GraphQL schema definition could benefit from the access to a type config I propose with this PR that we make the config field public by removing the underscore. From `_typeConfig` to `typeConfig`, `_scalarConfig` to `scalarConfig`, `_enumConfig` to `enumConfig`, and etc.

Currently this PR does not actually rename anything, instead it opts to add a getter alias which just returns the config with an underscore. This has the nice side effect of also disallowing assignment to the property as `type.typeConfig = null` would be illegal.