TypeInfo.getEnumValue() and EnumType.getValue(name)
This adds two new APIs to the schema utilities. One is for `TypeInfo` which is typically used alongside an AST visitor - now you can call `getEnumValue()` to get the GraphQLEnumValue instance referred to by the EnumValue ast node under visit.

Also, `EnumType.getValue(name)` accepts the name of an enum value and returns the GraphQLEnumValue instance associated with it. Beats running a `getValues().find(...)`.