RFC: More flow-friendly thenable checking.
This replaces the isThenable boolean check with a "getIfPromise" function. It's essentially the same, but rather than returning boolean, it's the identify function when it would have returned true, and returns void otherwise. That lets us give it a better type signature that Flow can do more with.

Open to feedback on this one, and open to bikeshedding on the name of this new function.