Use console.warn in name validation
In fe346190671c4ad0d666e8eef we degraded the hard error (ie. a `throw`)
for non-compliant field names (ie. starting with `__`) to a
`console.error`.

We've found, however, that certain CI systems will treat the use of
`console.error` as a cause for failure, and our intent here is not to
block what would otherwise be a valid test run.

Switch to `console.warn` instead.

Eventually in a later release this will become a hard error, but for now
continuing to treat this like a "deprecation" and just warn is
appropriate.