Avoid "Error:" text in advisory warnings
This goes a step further than the parent commit, which downgrades
`console.error` to `console.warn` for cases that are really just
advisory "deprecation" warnings for non-compliant schemas.

The motivation in that commit was preventing CI failures stemming from
use of `console.error`.

That may not be enough though, because on some JS engines (notably
Chrome/Node), "Error: " still appears in the output thanks to our use of
`Error` objects to get stack traces. Depending on how the CI is set-up,
that could also be enough to spuriously fail the run.

So in this commit, we add a `formatWarning` helper that takes the
`Error` object and grooms it for use as a human-readable warning
message.

Added tests to capture and verify the different engine behaviors.