Note that failing dunderscore name validation will become a hard error
To adequately serve as a deprecation warning, the message should provide
advance notice that it non-compliance will eventually become a hard
error. I haven't put a version number in there yet because I don't know
which version it will happen in. Once we make that call we should make
the message more precise.