Add constructor to GraphQLError class flow type
Since GraphQLError is declared then defined separately, Flow might infer that the constructor is not being used correctly via wrong arity, or might not detect improper use if an incorrect type is provided. This fixes that by adding a contructor to the declared type.

Also bumped the version of flow to ensure this change passes both the existing and most recent flow version