Flow v0.47.0 finds 3 errors in master. This is important to fix, because the graphql npm module ships with types, so anyone who uses the graphql npm module will get errors when upgrading to Flow v0.47.0. The new errors [look like this](https://gist.github.com/85d7f03510760dd223e968d2e92e7712).
When investigating, I found that `typeFromAST()` was doing something a little
odd. I believe it was trying to declare itself as an overloaded function, but
instead it was declaring an overloaded function and then shadowing it with the
exported function.

When it's properly fixed, it exposes 26 errors, which is a little much for me
to fix right now. So instead, I'm sabotaging its type completely.