Improve errors when receiving bad introspection data
I've seen a few examples of passing bad introspection data to buildClientSchema() result in a hard to understand internal error. This instead prints a more understandable error which includes a print out of the bad value in question. Hopefully this aids debugging.

This also adds a comment to the description of this function recommending that the "errors" field of a server response be checked before continuing to avoid partial results.