Improve introspection types + new getIntrospectionQuery()
This adds a new function `getIntrospectionQuery()` which allows for some minor configuration over the resulting query text: to exclude descriptions if your use case does not require them.

It also updates the flow types to remove exact types for the kinds of `__Type` since they will have other fields that would be `null`.