Improve dedent utility tag function used for tests
Changed the behavior of the function to not have the undocumented side
effect of converting the string to its escaped (raw) form any more,
and to also consider tabs in addition to spaces.

Also added unit tests for the dedent function and adapted existing
tests that were depending on the old behavior of the function.