Allow buildSchema() to take options.
buildSchema() is essentially just short hand for buildASTSchema(parse()), however since both of these composed functions accept options, it's nice to allow buildSchema() to accept options as well - it accepts the combined set of options which it passes to both functions.

This also names and exports the options for buildASTSchema() such that they can be used in the same way by third party code.